Make pull down menus appear briefly when switched to new screen
The pull down menus used to be invisible, and appear only when you moved
the mouse near them, so you kind of had to know they were there to use
them.

This change makes it so that when you switch to a new screen for the
first time, the menu briefly appears so the user has a chance to find
out it exists without being omniscient or just stumbling upon it by
luck.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>