Create initial window at 80% size of screen to fix aspect ratio.
If you ask SDL2 to create a window of exactly the size of the screen,
it appears to shrink your request down a little bit to make room for
the window title bar and borders, which messes with the aspect ratio
of the area you can draw in. If we shrink our request down so the borders
and title bar can fit, then our initial aspect ratio isn't messed up.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>