Fix bug in update_assets_from_launcher.sh if XDG_DATA_HOME set
If XDG_DATA_HOME was set, SNIS_DB_DIR was set to the same value, but
should have had "/space-nerds-in-space" appended to it.
Because snis_client.c could not see snis_downloading.txt it would not
display the download processing text or show the restart snis_client
when downloaded files were complete.

This fixes that issue.

Signed-off-by: vpelss@gmail.com
Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>