Do not load normal/tangent/bitangent debug vertices to GPU by default
In graph_dev_opengl, there is code to draw debug normal, tangent, and
bitangent lines for all triangle meshes, enabled via a special graphics
engine menu accessed via Ctrl-G, (but historically, via the SCROLL LOCK key).
For that to work, vertices for normal, tangent, and bitangent lines for
every vertex of every triangle mesh were uploaded to the GPU.

Now, this is hidden behind "#define DEBUG_NORMALS 0", so it is inaccessible
unless you compile it in, because we shouldn't be uploading a crap ton of
debug info to the GPU by default.

(To be fair, we weren't uploading it every frame, but only once for each triangle
mesh, so this is more about saving a little GPU memory than about performance.)

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>