lua-api: allow looking up objects by name.
Sometimes you have an NPC ship that is key to the plot.
Sometimes, that NPC ship gets destroyed.  The script needs
a way to know that this has happened, and do something about it.
This is a bit of a band-aid for now.  Maybe there should be some sort of
"plot armor" attribute.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>