Operation monkeywrench: change message delivery order
Was delivering the "encrypted" message first, then the other
messages directing the player what to do, but in playtesting,
players ignored everything but the encrypted message until it
was solved, spoiling the rest of the mission.  The original
idea behind that order was that it would take some time for the
encrypted message to be decoded (which it did), and I wanted to
give the players a head start on it.  That didn't work out well.

So now, deliver the encrypted message later on after they have
been given orders to destroy the Aurore.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>