Improve docking system
Require the player to meet minimum alignment requirement with docking
ports and to meet maximum relative velocity requirement with docking
ports.  These are tweakable via server side variables max_docking_vdiff
and min_docking_alignment.

The docking alignment is a dot product of a unit vector of the player's
direction with a unit vector of the docking port orientation.  Default
required aligment is 0.85.  Set to min_docking_alignment to 0 to disable
this requirement.

The max relative velocity is compared to the magnitude of the docking port
velocity vector minus the player's velocity vector.  The default max is 2.0.
This can be altered via the server side tweakable variable max_docking_vdiff.
Set it to some large value (e.g. 999.0) to disable this requirement.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>