Improve utilization of starbase models
When generating the universe, generate the starbases, then add the docking ports to the starbases
instead of adding the docking ports to the starbases as we go. That way the starbases get consecutive
IDs, and we don't skip over some of the starbase models (which are chosen by ID % number of models).

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>